vector: more basic functions
