+++ packages/vector/src/Vector.resi
@@ -20,6 +20,16 @@ let getExn: (t<'a>, int) => 'a
 let getUnsafe: (t<'a>, int) => 'a
 
 @genType
+let getByU: (t<'a>, (. 'a) => bool) => option<'a>
+@genType
+let getBy: (t<'a>, 'a => bool) => option<'a>
+
+@genType
+let getIndexByU: (t<'a>, (. 'a) => bool) => option<int>
+@genType
+let getIndexBy: (t<'a>, 'a => bool) => option<int>
+
+@genType
 let set: (t<'a>, int, 'a) => option<t<'a>>
 @genType
 let setExn: (t<'a>, int, 'a) => t<'a>
@@ -32,9 +42,20 @@ let reduceU: (t<'a>, 'b, (. 'b, 'a) => 'b) => 'b
 let reduce: (t<'a>, 'b, ('b, 'a) => 'b) => 'b
 
 @genType
+let reduceWithIndexU: (t<'a>, 'b, (. 'b, 'a, int) => 'b) => 'b
+@genType
+let reduceWithIndex: (t<'a>, 'b, ('b, 'a, int) => 'b) => 'b
+
+@genType
 let mapU: (t<'a>, (. 'a) => 'b) => t<'b>
 @genType
 let map: (t<'a>, 'a => 'b) => t<'b>
+
+@getType
+let mapWithIndexU: (t<'a>, (. 'a, int) => 'b) => t<'b>
+
+@genType
+let mapWithIndex: (t<'a>, ('a, int) => 'b) => t<'b>
 
 @genType
 let keepU: (t<'a>, (. 'a) => bool) => t<'a>
@@ -47,9 +68,19 @@ let keepMapU: (t<'a>, (. 'a) => option<'b>) => t<'b>
 let keepMap: (t<'a>, 'a => option<'b>) => t<'b>
 
 @genType
+let keepWithIndexU: (t<'a>, (. 'a, int) => bool) => t<'a>
+@genType
+let keepWithIndex: (t<'a>, ('a, int) => bool) => t<'a>
+
+@genType
 let forEachU: (t<'a>, (. 'a) => unit) => unit
 @genType
 let forEach: (t<'a>, 'a => unit) => unit
+
+@genType
+let forEachWithIndexU: (t<'a>, (. 'a, int) => unit) => unit
+@genType
+let forEachWithIndex: (t<'a>, ('a, int) => unit) => unit
 
 @genType
 let someU: (t<'a>, (. 'a) => bool) => bool
